move OnLoadNewsListListener and OnLoadNewsDetailListener
Interface is used to develop specifications and don't need care for the
specific content of its implementations class because of its
reusability.Therefore,we should move the inner interface of
NewsModelImpl.
